+++ ci.sh
@@ -2,6 +2,7 @@
 
 if [ ! -f .env ];then cp .env.example .env; fi
 if [ ! -f app.prod.ini ];then cp gogs/app.ini gogs/app.prod.ini; fi
+if [ ! -f registry/config.yml ];then cp registry/config.default.yml registry/config.yml; fi
 
 cd webhooks
 
@@ -12,14 +13,13 @@ if [ ! -f update.sh ];then cp update.example.sh update.sh; fi
 cd ../
 
 if [ "$1" = "production" ];then echo "You are in Production, Please exec\n"; echo "$ docker-compose -f docker-compose.prod.yml up -d\n"; exit 0; fi
+if [ "$1" = "swarm" ];then echo -e "\nYou are in Swarm mode, Please exec\n\n$ docker swarm init\n$ docker stack deploy -c docker-stack.yml ci\n"; exit 0; fi
 
-docker-compose up -d mysql
+command -v docker-compose > /dev/null 2>&1
 
-docker-compose exec mysql mysql -uroot -pmytest
+if [ $? -ne 0 ];then echo "docker-compose not install"; exit 1; fi
 
-# create database gogs;
-# exit
-# docker-compose up -d
-
-echo "You are in Development, Please exec\n"
-echo "$ docker-compose up -d\n"
+echo
+echo "You are in Local Development,please exec"
+echo "$ docker-compose up -d"
+echo