+++ Changelog.md
@@ -1,3 +1,3 @@
-`Docker Compose` `1.17.1` `3.4`
+* `Docker Compose` `1.17.1` `3.4`
 
-coding.net 0.8.1 出现问题未能解决
+* coding.net 0.8.1 出现问题未能解决